discord_quran_bot/requirements.txt
discord.py
yt-dlp
gunicorn
flake8
pre-commit
pytest
